UINode Cleanup
A lot has changed in this class.  Some ideas are still be worked with.  A big problem that needs to be resolved is how a UINode is run on top of another and still gives a exit event to the other node even though they know nothing of each other.