Changed std::map to std::list
The problem with map is that it would not allow for the same UINode to be called from itself or present at the same time because the identifier must be unique.  The list fixes the problem.  Since we only use push_back, the search speed of map is no longer needed.